backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.2.2
torchvision==0.17.2
pillow==10.1.0
pandas==2.1.3
numpy==1.26.2
opencv-python==4.9.0.80 